Add @license GPL
coxmichael's original license on userstyles.org was 'CC BY - Creative
Commons Attribution' https://userstyles.org/styles/154282/theguardian

But openuserjs.org requires an OSI-approved license.

According to this page https://opensource.stackexchange.com/a/5022
the GPL should be compatible with CC BY.

But if coxmichael has any issue with this license, I would of course be
happy to change it.